cast index to double

+++ src/main/java/de/dennisguse/opentracks/services/announcement/VoiceAnnouncementManager.java
@@ -183,7 +183,7 @@ public class VoiceAnnouncementManager implements SharedPreferences.OnSharedPrefe
 
             Distance distance = trackStatistics.getTotalDistance();
 
-            int index = (int) (distance.dividedBy(distanceFrequency));
+            int index = (double) (distance.dividedBy(distanceFrequency));
             nextTotalDistance = distanceFrequency.multipliedBy(index + 1);
         }
